Move meshMoctree to new MoctMesher.js
